Remove use of RepairSite class
Repair sites are on the tiled map
Robots check the floor layer to see if they are one a special field
