+++ src/main/java/inf112/roborally/game/board/BoardCreator.java
@@ -5,8 +5,6 @@ import com.badlogic.gdx.maps.tiled.TiledMapRenderer;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
-import inf112.roborally.game.Main;
-import inf112.roborally.game.RoboRallyGame;
 
 public abstract class BoardCreator {
