+++ src/main/java/inf112/roborally/game/board/VaultBoard.java
@@ -2,7 +2,6 @@ package inf112.roborally.game.board;
 
 import inf112.roborally.game.RoboRallyGame;
 import inf112.roborally.game.objects.Flag;
-import inf112.roborally.game.objects.RepairSite;
 
 public class VaultBoard extends Board {
 
@@ -10,7 +9,6 @@ public class VaultBoard extends Board {
         flags.add(new Flag(13, 6, 1));
         flags.add(new Flag(5, 8, 2));
         flags.add(new Flag(11, 4, 3));
-        repairSites.add(new RepairSite(5, 2));
         createBoard(RoboRallyGame.VAULT);
         findLasers();
     }
